% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaky_current.R
\name{flaky_current}
\alias{flaky_current}
\title{Return current session context}
\source{
<https://docs.snowflake.net/manuals/sql-reference/functions-context.html>
}
\usage{
flaky_current(current_set, con = getOption("flaky.con"))
}
\arguments{
\item{current_set}{Optional character vector that contains the base of context variable. When missing all variables are returned.}

\item{con}{A connection object. Typically generated by [flaky_connect()].}
}
\description{
Queries for context variables related to the current Snowflake session. Examples are ROLE, CLIENT, etc.
}
\details{
The boilerplate parts of the base name are added by the function. For example `"region"` will become `CURRENT_REGION()` and `"role"` will become `CURRENT_ROLE()`.
}
\examples{
/dontrun{
# See all variables
flaky_current()

# See one variable
flaky_current("warehouse")

# See a few variables
flaky_current(c("role", "schema"))
}
}
