% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaky_show.R
\name{flaky_show}
\alias{flaky_show}
\title{Show Command}
\usage{
flaky_show(type, pattern, n = 50, con = getOption("flaky.con"))
}
\arguments{
\item{type}{The type of thing to show.}

\item{pattern}{Pattern to filter results. Uses `LIKE` style filtering.}

\item{n}{Number of rows to return. Use `-1` to show all rows.}

\item{con}{A connection object. Typically generated by [flaky_connect()].}
}
\value{
Tibble of results
}
\description{
Call the the various `show` commands and return results as a tibble.
}
\examples{
\dontrun{
# Show 50 warehouses
flaky_show("warehouses")

# Show all functions
flaky_show("functions", n = -1)

# Show the azure regions
flaky_show("regions", "AZURE\%")
}
}
