% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaky_show.R
\name{flaky_show_ls}
\alias{flaky_show_ls}
\title{List possible show commands}
\usage{
flaky_show_ls(pattern)
}
\arguments{
\item{pattern}{Pattern to detect in name of show command. Uses [stringr::str_detect()].}
}
\value{
tibble
}
\description{
Parses Snowflake docs for show commands and provides as a tibble.
}
\examples{
\dontrun{
 # Return all shows
 flaky_show_ls()

 # Find function ones
 flaky_show_ls("fun")

 # Combine with flaky_show
 flaky_show_ls() \%>\% slice(1) \%>\% pull(name) \%>\% flaky_show()
}
}
