% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaky_use.R
\name{flaky_use}
\alias{flaky_use}
\title{Use Snowflake objects}
\usage{
flaky_use(object, name, con = getOption("flaky.con"), db_name)
}
\arguments{
\item{object}{The object type to use.}

\item{name}{Name of the desired object.}

\item{con}{Connection object to apply the use command towards.}

\item{db_name}{If changing schema to a different database, the desired database.}
}
\value{

}
\description{
Change active Snowflake objects like user, warehouse, database, etc.
}
