% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaky_use.R
\name{flaky_use_ls}
\alias{flaky_use_ls}
\title{List possible use commands}
\usage{
flaky_use_ls(pattern)
}
\arguments{
\item{pattern}{Pattern to detect in name of use command. Uses [stringr::str_detect()].}
}
\value{
tibble
}
\description{
Parses Snowflake docs for use commands and provides as a tibble.
}
\examples{
\dontrun{
 # Return all shows
 flaky_use_ls()

 # Find function ones
 flaky_use_ls("fun")
}
}
