% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaky_write.R
\name{flaky_write}
\alias{flaky_write}
\title{Write data to Snowflake}
\usage{
flaky_write(.data, table_schema, table_name, ..., con = getOption("flaky.con"))
}
\arguments{
\item{.data}{A tbl. The data we wish to write.}

\item{table_schema}{Name of target schema.}

\item{table_name}{Name of target table}

\item{...}{Additional parameters to pass on to [DBI::dbWriteTable()]. append, overwrite, etc.}

\item{con}{A connection object. Typically generated by [flaky_connect()].}
}
\description{
Move a local data frame to Snowflake. Ensures table and column names are upper case.
}
\examples{
\dontrun{
# Write to rdata.mtcars
mtcars \%>\% flaky_write("rdata", "mtcars")
}
}
