% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-snowflake.R
\name{sql_translate_env.Snowflake}
\alias{sql_translate_env.Snowflake}
\title{Snowflake Backend for dbplyr}
\usage{
sql_translate_env.Snowflake(con)
}
\arguments{
\item{con}{Connection object to use}
}
\description{
Snowflake Backend for dbplyr
}
