% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-snowflake.R
\name{sql_translation.Snowflake}
\alias{sql_translation.Snowflake}
\title{Snowflake Backend for dbplyr}
\usage{
\method{sql_translation}{Snowflake}(con)
}
\arguments{
\item{con}{Connection object to use}
}
\description{
Snowflake Backend for dbplyr
}
